Make accounts date helper an interface
